Fix debian build
